+++ src/Imagick/Imagine.php
@@ -27,7 +27,7 @@ use Imagine\Image\Palette\RGB;
 /**
  * Imagine implementation using the Imagick PHP extension.
  */
-final class Imagine extends AbstractImagine
+class Imagine extends AbstractImagine
 {
     /**
      * @throws RuntimeException