+++ src/Gd/Image.php
@@ -160,9 +160,6 @@ final class Image extends AbstractImage
         }
 
         $size = $image->getSize();
-        if (!$this->getSize()->contains($size, $start)) {
-            throw new OutOfBoundsException('Cannot paste image of the given size at the specified position, as it moves outside of the current image\'s box');
-        }
 
         if ($alpha === 100) {
             imagealphablending($this->resource, true);