+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ### NEXT (YYYY-MM-DD)
   * The `coalesce` method of `LayerInterface` instances now returns the LayerInterface itself (@mlocati)  
     **BREAKING CHANGE** if you have your own `LayerInterface` implementation, it now must return `$this`
+  * The `__toString` method has been added to `ColorInterface` since all its implementations have it (@mlocati)  
+    **BREAKING CHANGE** if you have your own `ColorInterface` implementation, it now must implement `__toString`
 
 ### 1.0.0-alpha1 (2018-08-28)
   * Imagine is now tested under Windows too (@mlocati)